refactor: renove portfolio with typescript
