--- src/components/custom/contact.module.css
@@ -1,159 +1,0 @@
-.contact {
-  position: relative;
-  width: 100%;
-  padding: 3em 0;
-}
-
-.cover {
-  position: relative;
-  padding: 1em 0 0 4em;
-}
-
-.title {
-  position: relative;
-  font-weight: bold;
-  font-size: 2em;
-  margin-bottom: 1em;
-  padding-bottom: 1em;
-  line-height: 0;
-  color: var(--color-turqui);
-}
-
-.title::after {
-  content: '';
-  position: absolute;
-  display: block;
-  width: 50px;
-  height: 3px;
-  background: var(--color-sky);
-  bottom: 0;
-  left: 0;
-  text-indent: 3em;
-}
-
-.section {
-  padding: 2em 0;
-}
-
-.section h3 {
-  font-weight: 700;
-  font-size: 1.4em;
-  color: var(--color-turqui);
-  text-align: center;
-}
-
-.container {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: space-evenly;
-  align-items: center;
-}
-
-.group {
-  width: 100%;
-}
-
-.info {
-  width: 80%;
-  padding: 2em;
-  box-shadow: 0 0 6px var(--opacity-black);
-}
-.info h3 {
-  font-weight: 700;
-  font-size: 1.2em;
-  color: var(--color-turqui);
-  padding-bottom: 1em;
-}
-
-.info i {
-  color: var(--color-sky);
-  display: flex;
-  float: left;
-  justify-content: center;
-  align-items: center;
-  background: #dff3fc;
-  font-size: 20px;
-  width: 44px;
-  height: 44px;
-  border-radius: 50px;
-  transition: all 0.3s ease-in-out;
-}
-
-.info h4 {
-  padding: 0 0 0 60px;
-  font-size: 1.2em;
-  font-weight: 600;
-  margin-bottom: 5px;
-  color: var(--color-turqui);
-}
-
-.info p {
-  padding: 0 0 10px 60px;
-  margin-bottom: 20px;
-  font-size: 0.9em;
-  color: var(--color-turqui);
-}
-
-.form {
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
-  grid-template-rows: repeat(auto-fit, minmax(250px, 1fr));
-  padding: 1em;
-}
-
-.data {
-  padding: 2em;
-  border: none;
-}
-
-.data label {
-  color: var(--color-turqui);
-  font-weight: 500;
-}
-
-.data input,
-textarea {
-  width: 100%;
-  padding: 1em;
-  margin: 0 0 1em 0;
-  border: 0.5px solid transparent;
-  border-right: 3px solid var(--color-turqui);
-  border-bottom: 1px solid var(--color-sky);
-  outline: none;
-  text-indent: 0.5em;
-  color: var(--color-turqui);
-  resize: none;
-  font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
-}
-
-.data textarea {
-  height: 110px;
-}
-
-.data input:hover,
-textarea:hover {
-  border-color: var(--color-sky);
-  background-color: #dff3fc;
-}
-
-.data input:focus,
-textarea:focus {
-  border-color: var(--color-sky);
-  background-color: #dff3fc;
-}
-
-.data button {
-  background-color: var(--color-water);
-  border: 1px solid var(--color-white);
-  color: var(--color-white);
-  padding: 1em;
-  font-size: 1em;
-  border-radius: 5%;
-}
-
-.data button:hover {
-  background-color: var(--color-sky);
-  color: var(--color-white);
-  cursor: pointer;
-  border: 1px solid var(--color-sky);
-}
