--- src/components/custom/portfolio.module.css
@@ -1,95 +1,0 @@
-.portfolio {
-  display: flex;
-  flex-direction: column;
-  padding: 4em;
-}
-
-.title {
-  font-size: 2em;
-  font-weight: bold;
-  margin-bottom: 1em;
-  padding-bottom: 1em;
-  position: relative;
-  color: var(--color-turqui);
-  line-height: 0;
-}
-
-.title::after {
-  content: '';
-  position: absolute;
-  display: block;
-  width: 50px;
-  height: 3px;
-  background-color: var(--color-sky);
-  bottom: 0;
-  left: 0;
-  text-indent: 3em;
-}
-
-.container {
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
-  grid-template-rows: repeat(auto-fit, minmax(250px, 1fr));
-  grid-gap: 2em;
-  width: 100%;
-  justify-content: center;
-  align-content: center;
-}
-
-.item {
-  position: relative;
-  width: 100%;
-}
-
-.item h4 {
-  position: relative;
-  text-align: center;
-  line-height: 0;
-  font-size: 1em;
-  color: var(--color-turqui);
-  text-transform: uppercase;
-  z-index: 20;
-}
-
-.item:hover .links {
-  opacity: 1;
-  bottom: 0;
-  z-index: 30;
-}
-.image {
-  position: relative;
-  width: 100%;
-  height: auto;
-  object-fit: cover;
-  z-index: 20;
-}
-.item:hover .image,
-.item:hover h4 {
-  opacity: 0.5;
-}
-
-.links {
-  position: absolute;
-  display: flex;
-  z-index: 3;
-  right: 0;
-  left: 0;
-  bottom: 5px;
-  opacity: 0;
-  transition: all ease-in-out 0.3s;
-  justify-content: center;
-  z-index: 0;
-}
-
-.links a {
-  font-size: 28px;
-  text-align: center;
-  background-color: var(--color-lightblue);
-  transition: 0.3s;
-  width: 50%;
-  cursor: pointer;
-  border-left: 1px solid var(--color-water);
-}
-.links a:hover {
-  background: var(--color-water);
-}
