+++ src/components/layout/Header/header.module.css
@@ -1,45 +1,34 @@
-.header {
+/* --------------------------------- sidenav -------------------------------- */
+
+.sidenav {
   background: var(--color-ebony);
-  position: fixed;
   overflow-y: auto;
   transition: all ease 1s;
   transition: all 1s;
-  width: 220px;
-  top: 0;
-  left: 0;
-  bottom: 0;
-  padding: 0 50px;
-  z-index: 9000;
+  width: 355px;
+  position: relative;
+  height: 100%;
 }
 
-.mobile {
-  background: var(--color-ebony);
-  position: fixed;
-  overflow-y: auto;
-  transition: all ease 1.5s;
-  transition: all 1.5s;
-  transition-delay: 0.4s;
-  width: 220px;
-  top: 0;
-  left: -500px;
-  bottom: 0;
-  padding: 0 50px;
-  z-index: 9000;
-}
-
-.header::-webkit-scrollbar {
+.sidenav::-webkit-scrollbar {
   background-color: var(--color-black);
   width: 0.5em;
 }
 
-.header::-webkit-scrollbar-thumb {
+.sidenav::-webkit-scrollbar-thumb {
   background-color: var(--color-blue);
   border-radius: 0.5em;
 }
 
+/* -------------------------------- content ------------------------------- */
+
 .container {
+  background: var(--color-ebony);
+  position: fixed;
   display: flex;
   flex-direction: column;
+  height: 100vh;
+  padding: 6rem 4rem 0 4rem;
 }
 
 .image {
@@ -69,15 +58,24 @@
   text-decoration: none;
 }
 
+/* --------------------------------- social --------------------------------- */
+
+.group {
+  padding: 1rem 0;
+}
+
 .social {
   text-align: center;
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
 }
 
 .social a {
   background: var(--color-lobezno);
   color: var(--color-white);
   display: inline-block;
-  padding: 0.5em;
+  padding: 0 0.5em;
   margin: 1em 0.1em 0 0.1em;
   border-radius: 50%;
   font-size: 1.2em;
@@ -93,8 +91,10 @@
   text-decoration: none;
 }
 
+/* ---------------------------------- menu ---------------------------------- */
+
 .menu {
-  padding-top: 1.8em;
+  padding-top: 2rem;
 }
 
 .menu * {
@@ -112,7 +112,7 @@
   display: flex;
   align-items: center;
   color: var(--color-pearl);
-  padding: 1em;
+  padding: 1em 0;
   margin-bottom: 0.5em;
   transition: 0.3s;
   font-size: 1em;
@@ -138,6 +138,23 @@
 .menu li:hover > .link i {
   color: var(--color-sky);
 }
+
+/* 
+.mobile {
+  background: var(--color-ebony);
+  position: fixed;
+  overflow-y: auto;
+  transition: all ease 1.5s;
+  transition: all 1.5s;
+  transition-delay: 0.4s;
+  width: 220px;
+  top: 0;
+  left: -500px;
+  bottom: 0;
+  padding: 0 50px;
+  z-index: 9000;
+}
+
 
 .toggle {
   position: fixed;
@@ -169,4 +186,4 @@
   padding: 0;
   border: 0;
   color: var(--color-sky);
-}
+} */
