--- src/css/index.css
@@ -1,106 +1,0 @@
-@import 'colors.css';
-
-@import url('https://fonts.googleapis.com/css2?family=Alata&display=swap');
-
-body {
-  margin: 0;
-  font-family: 'Open Sans', -apple-system, BlinkMacSystemFont, 'Segoe UI',
-    'Roboto', 'Oxygen', 'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans',
-    'Helvetica Neue', sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-  color: #272829;
-}
-
-code {
-  font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
-    monospace;
-}
-
-a {
-  color: var(--color-sky);
-}
-
-a:hover {
-  color: var(--color-water);
-  text-decoration: none;
-}
-
-h1,
-h2,
-h3,
-h4,
-h5,
-h6 {
-  font-family: 'Raleway', sans-serif;
-}
-
-.html,
-.css,
-.js,
-.react,
-.node,
-.mongo,
-.mysql,
-.git,
-.github,
-.heroku,
-.firebase {
-  font-size: 1.5em;
-  padding-right: 5px;
-}
-
-.excel,
-.access,
-.vba,
-.powerbi,
-.tableau,
-.jamovi {
-  width: 1.5em;
-  height: 1.5em;
-  padding-right: 5px;
-}
-
-.html {
-  color: #fc490b;
-}
-
-.css {
-  color: #2299f8;
-}
-
-.js {
-  color: #f7df1e;
-}
-
-.react {
-  color: #61dafb;
-}
-
-.node {
-  color: #689f63;
-}
-
-.mongo {
-  color: #10aa50;
-}
-
-.mysql {
-  color: #007699;
-}
-
-.git {
-  color: #f05133;
-}
-
-.github {
-  color: #000000;
-}
-
-.heroku {
-  color: #430098;
-}
-
-.firebase {
-  color: #ffcc2b;
-}
