--- src/pages/main.module.css
@@ -1,313 +1,0 @@
-.all {
-  width: 100%;
-  position: relative;
-  margin-left: 0;
-  transition: all ease-in-out 1s;
-  transition: all 1s;
-}
-
-.margin {
-  position: relative;
-  margin-left: 320px;
-  transition: all ease-in-out 1s;
-  transition: all 1s;
-  transition-delay: 0.4s;
-}
-
-.hero {
-  display: flex;
-  flex-direction: column;
-  position: relative;
-  width: 100%;
-  height: 100vh;
-}
-
-.container {
-  display: flex;
-  flex-direction: column;
-  position: relative;
-  flex-wrap: wrap;
-  align-self: center;
-  padding: 3em 0 0 0;
-  z-index: 20;
-  width: 80%;
-  height: 100vh;
-}
-
-.hero h1 {
-  font-size: 4em;
-  font-weight: 700;
-  margin: 0;
-  color: var(--color-white);
-  font-family: 'Alata', sans-serif;
-}
-
-.hero h3 {
-  font-family: 'Alata', sans-serif;
-}
-
-.hero p {
-  color: var(--color-white);
-  font-size: 1.5em;
-  font-family: 'Poppins', sans-serif;
-}
-
-.hero p span {
-  color: #fff;
-  letter-spacing: 1px;
-  border-bottom: 3px solid #149ddd;
-}
-
-.about {
-  display: flex;
-  justify-content: center;
-}
-
-.containerAbout {
-  padding: 60px 0;
-  overflow: hidden;
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-}
-
-.section {
-  padding-bottom: 30px;
-  width: 90%;
-  align-self: center;
-}
-
-.section h2 {
-  font-size: 32px;
-  font-weight: bold;
-  margin-bottom: 20px;
-  padding-bottom: 20px;
-  position: relative;
-  color: #173b6c;
-}
-
-.section h2::after {
-  content: '';
-  position: absolute;
-  display: block;
-  width: 50px;
-  height: 3px;
-  background: #149ddd;
-  bottom: 0;
-  left: 0;
-}
-
-.section p {
-  margin-bottom: 0;
-  line-height: 2.5;
-  text-align: justify;
-  text-indent: 3em;
-}
-.figure {
-  max-width: 350px;
-  width: 100%;
-  padding: 0 0 2em 2em;
-  margin: 0;
-  height: auto;
-  float: right;
-}
-
-.imagen {
-  width: 100%;
-  margin: 0;
-  padding: 0;
-  box-shadow: 6px 6px 6px #0008;
-}
-
-.skills {
-  padding-top: 1em;
-  clear: both;
-}
-
-.content {
-  text-align: center;
-}
-
-.content h3 {
-  font-weight: 700;
-  font-size: 1.6em;
-  color: var(--color-turqui);
-}
-
-.group {
-  position: relative;
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
-  grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
-  list-style: none;
-  padding: 0 2em 3em 2em;
-}
-
-.group li {
-  padding: 10px;
-  vertical-align: center;
-}
-
-.group i {
-  font-size: 1em;
-  padding-right: 2px;
-  color: #149ddd;
-}
-
-.content p:last-child {
-  margin-bottom: 0;
-}
-
-.align {
-  word-spacing: 0.5em;
-  letter-spacing: 2px;
-  vertical-align: text-top;
-  padding: 20px;
-}
-
-.document {
-  display: flex;
-  justify-content: center;
-  flex-wrap: wrap;
-}
-
-.cv {
-  margin: 0.5em;
-  padding: 0.5em;
-  border: 2px solid var(--color-sky);
-  border-radius: 15px;
-}
-
-.cv:hover {
-  background-color: #00bbff18;
-  color: white;
-}
-
-.resume {
-  display: flex;
-  flex-direction: column;
-  min-height: 100vh;
-  padding: 3em 0;
-}
-
-.containerResume {
-  width: 90%;
-  align-self: center;
-}
-.study {
-  width: 85%;
-  align-self: center;
-}
-
-.title {
-  font-size: 2em;
-  font-weight: bold;
-  margin-bottom: 1em;
-  padding-bottom: 1em;
-  position: relative;
-  color: var(--color-turqui);
-  line-height: 0;
-}
-
-.title::after {
-  content: '';
-  position: absolute;
-  display: block;
-  width: 50px;
-  height: 3px;
-  background: var(--color-sky);
-  bottom: 0;
-  left: 0;
-  text-indent: 3em;
-}
-
-.item {
-  padding: 0 0 2em 2em;
-  margin-top: -2px;
-  border-left: 2px solid var(--color-turqui);
-  position: relative;
-}
-
-.item h4 {
-  line-height: 1.5em;
-  font-size: 1em;
-  font-weight: 600;
-  text-transform: uppercase;
-  font-family: 'Poppins', sans-serif;
-  color: #173b6c;
-  margin-bottom: 0.5em;
-}
-
-.item h5 {
-  font-size: 1em;
-  margin: 0;
-  background: #eaf3fd;
-  padding: 0.5em 1em;
-  display: inline-block;
-  font-weight: 600;
-  border-radius: 10%;
-  line-height: 1.5em;
-}
-
-.item ul {
-  padding-left: 2em;
-  list-style-type: circle;
-}
-
-.item ul li {
-  padding-bottom: 1em;
-}
-
-.item:last-child {
-  padding-bottom: 0;
-}
-
-.red {
-  color: var(--color-red);
-  font-weight: 600;
-}
-
-.green {
-  color: var(--color-green);
-  font-weight: 600;
-}
-
-.item::before {
-  content: '';
-  position: absolute;
-  width: 1em;
-  height: 1em;
-  border-radius: 3em;
-  left: -0.5em;
-  top: 0;
-  background-color: var(--color-white);
-  border: 2px solid var(--color-turqui);
-}
-
-@media screen and (max-width: 700px) {
-  .all {
-    width: 100%;
-    position: relative;
-    margin-left: 0;
-    transition: all ease-in-out 1s;
-    transition: all 1s;
-  }
-
-  .margin {
-    position: relative;
-    margin-left: 0;
-    transition: all ease-in-out 1s;
-    transition: all 1s;
-    transition-delay: 0.4s;
-  }
-}
-
-@media screen and (max-width: 400px) {
-  .hero h1 {
-    font-size: 2.5em;
-  }
-
-  .hero p {
-    font-size: 1em;
-  }
-}
